Added script, tried to reformulate pipeline.
The script in question starts the parser on each machine in the cluster, and
then checks that it's working correctly.

The pipeline reformulation is a failed attempt to get the parsing to be done
across the entire cluster. No idea what's up with it but it insists on running
on a single machine ... slowly.