Integrated mlnsem, implemented convert to CNF
The integration with mlnsem requires modifying mlnsem and scala-logic -
in particular their build files need to be hacked up to allow local
deployment (to the versions specified in build.sbt), and certain methods
need to be de-privatized.

ConvertToCNF is inside trc1.mlnsem - it implements the whole procedure
using the scala-logic FOLExpression representation (except in the last
stage, where it adds the stuff in Distribute.scala - list mode
conjunctions and disjunctions).