boxer FOL, stage fix, code cleanup
decided to stop waiting around for a fix for the boxer drs parser, and
instead implemented a parser for the boxer FOL ouput. with this
implemented, I can now generate CNF expressions for sentences.

some calling tweaks allow negation of the transformed version at a
sensible point, and this lead to a code cleanup (removing unused
classes/objects, removing unused files, renaming packages).