made another serious pipeline change, and fixed up some stuff
serious pipeline change: figured out how to force the resolver to be run
in a task separate from tasks that call out to boxer. this, i hope, will
improve performance even more.

fixed up a map creation thing - it turns out that to make a frequency
count map via map and reduce, it's a good idea to use the Monoid type
class - it'll do the merge properly for you (if the monoids involved are
implemented right).

also cleaned up a whole bunch of imports. collections in resolution are
now definitely going to be the immutable ones, and some other cruft was
also removed. the big thing is that keeping the number of scalaz imports
down is a good idea - it makes compiling faster. no joke.