think I've fixed #7 - got command proc killing in
basically, it has to set up the command run, and then set up a future
bound on the completion of the process which will return the result. it
then waits on the future until either completion or timeout. if the
timeout does occur, the method (in GetFOL) can actually destroy the
process that timed out. using validations, the computation is then
failed.

I also cleaned things up so that these validations all go to
MaxTransform, allowing it to handle the printing of failures (outside of
Resolution, though).